pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.5.0",
    "fastapi>=0.125.0",
    "google-cloud-bigquery>=3.39.0",
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.38.0",
]
